[#159956003] Create tested end points for API
