--- app/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "rest-api",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "dependencies": {
-    "chai": "^4.1.2",
-    "express": "^4.16.3",
-    "supertest": "^3.1.0"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "coveralls": "^3.0.2",
-    "eslint": "^5.4.0",
-    "istanbul": "^0.4.5",
-    "mocha": "^5.2.0",
-    "nyc": "^12.0.2",
-    "rimraf": "^2.6.2"
-  },
-  "scripts": {
-    "test": "mocha"
-  },
-  "author": "",
-  "license": "ISC"
-}
