+++ package.json
@@ -1,27 +1,57 @@
 {
-  "name": "rest-api",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "dependencies": {
-    "chai": "^4.1.2",
-    "express": "^4.16.3",
-    "mocha": "^5.2.0",
-    "supertest": "^3.1.0"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "coveralls": "^3.0.2",
-    "eslint": "^5.4.0",
-    "nyc": "^12.0.2"
-  },
-  "scripts": {
-    "build": "babel app -d build",
-    "test": "nyc mocha build/test/test.js",
-    "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "start": "npm run build && mocha build/test/test.js"
-  },
-  "author": "",
-  "license": "ISC"
+    "name": "rest-api",
+    "version": "1.0.0",
+    "description": "",
+    "main": "app.js",
+    "dependencies": {
+        "body-parser": "^1.18.3",
+        "chai": "^4.1.2",
+        "express": "^4.16.3",
+        "mocha": "^5.2.0",
+        "supertest": "^3.1.0"
+    },
+    "devDependencies": {
+        "babel-cli": "^6.26.0",
+        "babel-preset-es2015": "^6.24.1",
+        "coveralls": "^3.0.2",
+        "eslint": "^5.4.0",
+        "nyc": "^12.0.2"
+    },
+    "scripts": {
+        "build": "babel server -d build",
+        "test": "nyc mocha build/test/test.js",
+        "coverage": "nyc report --reporter=text-lcov | coveralls",
+        "start": "npm run build && mocha build/test/test.js"
+    },
+    "author": "",
+    "license": "ISC",
+    "eslintConfig": {
+        "env": {
+            "browser": true,
+            "es6": true,
+            "node": true
+        },
+        "extends": "airbnb",
+        "parserOptions": {
+            "ecmaVersion": 2015
+        },
+        "rules": {
+            "indent": [
+                "error",
+                4
+            ],
+            "linebreak-style": [
+                "error",
+                "windows"
+            ],
+            "quotes": [
+                "error",
+                "single"
+            ],
+            "semi": [
+                "error",
+                "always"
+            ]
+        }
+    }
 }
